// src/api_main/models/MenuResponseModelListBaseResponseModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * BaseWebAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { MenuResponseModel } from './MenuResponseModel';
import {
    MenuResponseModelFromJSON,
    MenuResponseModelFromJSONTyped,
    MenuResponseModelToJSON,
    MenuResponseModelToJSONTyped,
} from './MenuResponseModel';

/**
 * 
 * @export
 * @interface MenuResponseModelListBaseResponseModel
 */
export interface MenuResponseModelListBaseResponseModel {
    /**
     * 
     * @type {boolean}
     * @memberof MenuResponseModelListBaseResponseModel
     */
    success?: boolean;
    /**
     * 
     * @type {string}
     * @memberof MenuResponseModelListBaseResponseModel
     */
    message?: string;
    /**
     * 
     * @type {number}
     * @memberof MenuResponseModelListBaseResponseModel
     */
    statusCode?: number;
    /**
     * 
     * @type {Array<MenuResponseModel>}
     * @memberof MenuResponseModelListBaseResponseModel
     */
    data?: Array<MenuResponseModel>;
}

/**
 * Check if a given object implements the MenuResponseModelListBaseResponseModel interface.
 */
export function instanceOfMenuResponseModelListBaseResponseModel(value: object): value is MenuResponseModelListBaseResponseModel {
    return true;
}

export function MenuResponseModelListBaseResponseModelFromJSON(json: any): MenuResponseModelListBaseResponseModel {
    return MenuResponseModelListBaseResponseModelFromJSONTyped(json, false);
}

export function MenuResponseModelListBaseResponseModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): MenuResponseModelListBaseResponseModel {
    if (json == null) {
        return json;
    }
    return {
        
        'success': json['success'] == null ? undefined : json['success'],
        'message': json['message'] == null ? undefined : json['message'],
        'statusCode': json['statusCode'] == null ? undefined : json['statusCode'],
        'data': json['data'] == null ? undefined : ((json['data'] as Array<any>).map(MenuResponseModelFromJSON)),
    };
}

export function MenuResponseModelListBaseResponseModelToJSON(json: any): MenuResponseModelListBaseResponseModel {
    return MenuResponseModelListBaseResponseModelToJSONTyped(json, false);
}

export function MenuResponseModelListBaseResponseModelToJSONTyped(value?: MenuResponseModelListBaseResponseModel | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'success': value['success'],
        'message': value['message'],
        'statusCode': value['statusCode'],
        'data': value['data'] == null ? undefined : ((value['data'] as Array<any>).map(MenuResponseModelToJSON)),
    };
}